resolve linter errors
