+++ client/.eslintrc.yml
@@ -17,4 +17,7 @@ rules: {
   "no-nested-ternary": "warn",
   "no-param-reassign": "off",
   "no-plusplus": "off",
+  "jsx-a11y/no-static-element-interactions": "warn",
+  "jsx-a11y/click-events-have-key-events": "warn",
+  "jsx-a11y/no-noninteractive-element-interactions": "warn",
 }
