+++ client/src/Components/BoardEntry.jsx
