+++ client/src/Components/Game/Powerups/PowerupButton.jsx
@@ -1,5 +1,3 @@
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React from 'react';
 import ImageDisplay from '../../Store/ImageDisplay';
 import '../../../Styles/Powerup.css';
