+++ client/src/Pages/Game.jsx
@@ -1,5 +1,3 @@
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { useEffect } from 'react';
 import '../Styles/Game.css';
 import Button from '../Components/Button';
